requirements.txt
Django==4.2.2
python-dotenv==1.0.0
djangorestframework==3.14.0
django-storages==1.13.2
django-cors-headers==4.0.0
requests==2.31.0
boto3==1.26.148
celery==5.3.0
django-redis==5.2.0
stripe==5.4.0
daphne==4.0.0
psycopg2-binary==2.9.6
django-rest-knox==4.2.0
resend==0.5.1
svix==1.4.12
